不要箇所を削除した
