+++ src/ExerciseComponents/TodoApp/TodoList/useTodoList.ts
@@ -1,5 +1,6 @@
 import { useCallback, useReducer } from "react";
 import { reducer, initialState } from "./TodoList.reducer";
+import { Todo } from "../Todo.type";
 
 export function useTodoList() {
   const [{ todoList }, dispatch] = useReducer(reducer, initialState);
@@ -31,11 +32,36 @@ export function useTodoList() {
     });
   }, []);
 
+  const filteredByStatus = (todoList: Todo[], status: string) => {
+    if (status === 'all') return [...todoList];
+
+    if (status === 'completed' || status === 'active') {
+      const isComplete = status === 'completed';
+      const statusResult = todoList.filter(
+        todo => todo.completed === isComplete
+      )
+      return statusResult;
+    }
+    else {
+      return [...todoList];
+    }
+  }
+
+  const filteredByQuery = (todoList: Todo[], query: string | undefined) => {
+    if (query === undefined) return [...todoList];
+
+    const queryResult = todoList.filter(
+      todo => (todo.title.includes(query)));
+    return queryResult;
+  }
+
   return {
     todoList,
     toggleTodo,
     createTodo,
     deleteTodo,
+    filteredByStatus,
+    filteredByQuery,
   };
 }
